+++ src/main/java/com/cjm721/overloaded/client/render/entity/RenderMultiBoots.java
@@ -1,7 +1,12 @@
 package com.cjm721.overloaded.client.render.entity;
 
+import com.cjm721.overloaded.Overloaded;
+import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.platform.GlStateManager;
+import com.mojang.blaze3d.vertex.IVertexBuilder;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.entity.model.BipedModel;
+import net.minecraft.client.renderer.model.IBakedModel;
 import net.minecraft.client.renderer.model.ModelResourceLocation;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.util.ResourceLocation;
@@ -12,48 +17,14 @@ import static com.cjm721.overloaded.Overloaded.MODID;
 
 public class RenderMultiBoots extends AbstractRenderMultiArmor {
 
+    public static RenderMultiBoots INSTANCE;
+
     public RenderMultiBoots(BipedModel baseModel) {
-        super(baseModel);
-//    this.bipedRightLeg.cubeList.clear();
-//    this.bipedLeftLeg.cubeList.clear();
-//
-//    ModelRenderOBJ rightBoot =
-//        new ModelRenderOBJ(this, new ModelResourceLocation(MODID + ":multi_right_boot", "armor"));
-//    ModelRenderOBJ leftBoot =
-//        new ModelRenderOBJ(this, new ModelResourceLocation(MODID + ":multi_left_boot", "armor"));
-//
-//    rightBoot.offsetY = 0.76F;
-//    rightBoot.offsetX = -0.03F;
-//    leftBoot.offsetY = 0.76F;
-//    leftBoot.offsetX = 0.03F;
+        super(baseModel,-10);
+        IBakedModel leftLeg = Minecraft.getInstance().getModelManager().getModel(new ResourceLocation(Overloaded.MODID, "item/armor/multi_left_boot"));
+        IBakedModel rightLeg = Minecraft.getInstance().getModelManager().getModel(new ResourceLocation(Overloaded.MODID, "item/armor/multi_right_boot"));
 
-//    leftBoot.scale = 1F / 14F;
-//    rightBoot.scale = 1F / 14F;
-//
-//    this.bipedRightLeg.addChild(rightBoot);
-//    this.bipedLeftLeg.addChild(leftBoot);
-    }
-
-//    @Override
-    public void render(
-            @Nullable LivingEntity entity,
-            float limbSwing,
-            float limbSwingAmount,
-            float ageInTicks,
-            float netHeadYaw,
-            float headPitch,
-            float scale) {
-//    super.setRotationAngles(
-//        entity, limbSwing, limbSwingAmount, ageInTicks, netHeadYaw, headPitch, scale);
-
-        GlStateManager.pushMatrix();
-        if (this.isSneak) {
-            GlStateManager.translatef(0.0F, 0.2F, 0.0F);
-        }
-
-//    this.bipedRightLeg.render(scale);
-//    this.bipedLeftLeg.render(scale);
-
-        GlStateManager.popMatrix();
+        this.bipedLeftLeg.addChild(new ModelRenderOBJ(this, leftLeg));
+        this.bipedRightLeg.addChild(new ModelRenderOBJ(this, rightLeg));
     }
 }