+++ src/main/java/com/cjm721/overloaded/client/render/entity/RenderMultiChestplate.java
@@ -4,6 +4,8 @@ import com.cjm721.overloaded.Overloaded;
 import com.mojang.blaze3d.matrix.MatrixStack;
 import com.mojang.blaze3d.vertex.IVertexBuilder;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.renderer.Quaternion;
+import net.minecraft.client.renderer.Vector3f;
 import net.minecraft.client.renderer.entity.model.BipedModel;
 import net.minecraft.client.renderer.model.IBakedModel;
 import net.minecraft.util.ResourceLocation;
@@ -22,13 +24,5 @@ public class RenderMultiChestplate extends AbstractRenderMultiArmor {
     this.bipedBody.addChild(new ModelRenderOBJ(this, body));
     this.bipedLeftArm.addChild(new ModelRenderOBJ(this, leftArm));
     this.bipedRightArm.addChild(new ModelRenderOBJ(this, rightArm));
-  }
-
-  @Override
-  public void render(MatrixStack stack, IVertexBuilder buffer, int light, int overlay, float r, float g, float b, float a) {
-    super.render(stack, buffer, light, overlay, r, g, b, a);
-    this.bipedBody.render(stack, buffer, light, overlay, r, g, b, a);
-    this.bipedLeftArm.render(stack, buffer, light, overlay, r, g, b, a);
-    this.bipedRightArm.render(stack, buffer, light, overlay, r, g, b, a);
   }
 }