+++ src/main/java/com/cjm721/overloaded/item/functional/armor/ItemMultiBoots.java
@@ -2,6 +2,7 @@ package com.cjm721.overloaded.item.functional.armor;
 
 import com.cjm721.overloaded.client.render.dynamic.ImageUtil;
 import com.cjm721.overloaded.client.render.entity.RenderMultiBoots;
+import com.cjm721.overloaded.client.render.entity.RenderMultiLeggings;
 import com.cjm721.overloaded.config.OverloadedConfig;
 import net.minecraft.client.renderer.entity.model.BipedModel;
 import net.minecraft.client.renderer.model.ModelResourceLocation;
@@ -35,4 +36,14 @@ public class ItemMultiBoots extends AbstractMultiArmor {
         new ResourceLocation(MODID, "textures/item/multi_boot.png"),
         OverloadedConfig.INSTANCE.textureResolutions.multiArmorResolution);
   }
+//
+//  @Nullable
+//  @Override
+//  public <A extends BipedModel<?>> A getArmorModel(LivingEntity entityLiving, ItemStack itemStack, EquipmentSlotType armorSlot, A baseModel) {
+//    if(RenderMultiBoots.INSTANCE == null) {
+//      RenderMultiBoots.INSTANCE = new RenderMultiBoots(baseModel);
+//    }
+//
+//    return (A) RenderMultiBoots.INSTANCE;
+//  }
 }