+++ src/main/java/com/cjm721/overloaded/item/functional/armor/ItemMultiChestplate.java
@@ -5,6 +5,7 @@ import com.cjm721.overloaded.client.render.entity.RenderMultiChestplate;
 import com.cjm721.overloaded.config.OverloadedConfig;
 import net.minecraft.client.renderer.entity.model.BipedModel;
 import net.minecraft.client.renderer.model.ModelResourceLocation;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.inventory.EquipmentSlotType;
 import net.minecraft.item.ItemStack;
@@ -25,15 +26,15 @@ public class ItemMultiChestplate extends AbstractMultiArmor {
     //        setTranslationKey("multi_chestplate");
   }
 
-  @Nullable
-  @Override
-  public <A extends BipedModel<?>> A getArmorModel(LivingEntity entityLiving, ItemStack itemStack, EquipmentSlotType armorSlot, A baseModel) {
-    if(RenderMultiChestplate.INSTANCE == null) {
-      RenderMultiChestplate.INSTANCE = new RenderMultiChestplate(baseModel);
-    }
-
-    return (A) RenderMultiChestplate.INSTANCE;
-  }
+//  @Nullable
+//  @Override
+//  public <A extends BipedModel<?>> A getArmorModel(LivingEntity entityLiving, ItemStack itemStack, EquipmentSlotType armorSlot, A baseModel) {
+//    if(RenderMultiChestplate.INSTANCE == null) {
+//      RenderMultiChestplate.INSTANCE = new RenderMultiChestplate(baseModel);
+//    }
+//
+//    return (A) RenderMultiChestplate.INSTANCE;
+//  }
 
   @OnlyIn(Dist.CLIENT)
   @Override