+++ src/main/java/com/cjm721/overloaded/item/functional/armor/ItemMultiLeggings.java
@@ -1,6 +1,7 @@
 package com.cjm721.overloaded.item.functional.armor;
 
 import com.cjm721.overloaded.client.render.dynamic.ImageUtil;
+import com.cjm721.overloaded.client.render.entity.RenderMultiHelmet;
 import com.cjm721.overloaded.client.render.entity.RenderMultiLeggings;
 import com.cjm721.overloaded.config.OverloadedConfig;
 import net.minecraft.client.renderer.entity.model.BipedModel;
@@ -38,4 +39,15 @@ public class ItemMultiLeggings extends AbstractMultiArmor {
         new ResourceLocation(MODID, "textures/item/multi_belt.png"),
         OverloadedConfig.INSTANCE.textureResolutions.multiArmorResolution);
   }
+
+//  @Nullable
+//  @Override
+//  public <A extends BipedModel<?>> A getArmorModel(LivingEntity entityLiving, ItemStack itemStack, EquipmentSlotType armorSlot, A baseModel) {
+//    if(RenderMultiLeggings.INSTANCE == null) {
+//      RenderMultiLeggings.INSTANCE = new RenderMultiLeggings(baseModel);
+//    }
+//
+//    return (A) RenderMultiLeggings.INSTANCE;
+//  }
+
 }